move valid_window and gcps to extractor
