+++ src/index_safe/create_index.py
@@ -11,8 +11,6 @@ from pathlib import Path
 from typing import Iterable, Union
 
 import boto3
-import numpy as np
-from osgeo import gdal
 from tqdm import tqdm
 
 
@@ -20,103 +18,6 @@ try:
     from index_safe import utils
 except ModuleNotFoundError:
     import utils
-
-gdal.UseExceptions()
-
-
-def compute_valid_window(index: int, burst: ET.Element) -> utils.Window:
-    """Written by Jason Ninneman for the ASF I&A team's burst extractor.
-    Using the information contained within a SAFE annotation XML burst
-    element, identify the window of a burst that contains valid data.
-
-    Args:
-        index: zero-indexed burst number to compute the valid window for
-        burst: <burst> element of annotation XML for the given index
-
-    Returns:
-        Row and column ranges for the valid data within a burst
-    """
-    # all offsets, even invalid offsets
-    offsets_range = utils.Offset(
-        np.array([int(val) for val in burst.find('firstValidSample').text.split()]),
-        np.array([int(val) for val in burst.find('lastValidSample').text.split()]),
-    )
-
-    # returns the indices of lines containing valid data
-    lines_with_valid_data = np.flatnonzero(offsets_range.stop - offsets_range.start)
-
-    # get first and last sample with valid data per line
-    # x-axis, range
-    valid_offsets_range = utils.Offset(
-        offsets_range.start[lines_with_valid_data].min(),
-        offsets_range.stop[lines_with_valid_data].max(),
-    )
-
-    # get the first and last line with valid data
-    # y-axis, azimuth
-    valid_offsets_azimuth = utils.Offset(
-        lines_with_valid_data.min(),
-        lines_with_valid_data.max(),
-    )
-
-    # x-length
-    length_range = valid_offsets_range.stop - valid_offsets_range.start
-    # y-length
-    length_azimuth = len(lines_with_valid_data)
-
-    valid_window = utils.Window(
-        valid_offsets_range.start,
-        valid_offsets_range.start + length_range,
-        valid_offsets_azimuth.start,
-        valid_offsets_azimuth.start + length_azimuth,
-    )
-
-    return valid_window
-
-
-def get_gcps_from_xml(annotation_xml: ET.Element) -> Iterable[utils.GeoControlPoint]:
-    """Get geolocation control points from annotation XML.
-
-    Args:
-        annotation_xml: root element of annotation XML
-
-    Returns:
-        List of geolocation control points
-    """
-    xml_gcps = annotation_xml.findall('.//{*}geolocationGridPoint')
-    gcps = []
-    for xml_gcp in xml_gcps:
-        pixel = int(xml_gcp.findtext('.//{*}pixel'))
-        line = int(xml_gcp.findtext('.//{*}line'))
-        longitude = round(float(xml_gcp.findtext('.//{*}longitude')), 7)
-        latitude = round(float(xml_gcp.findtext('.//{*}latitude')), 7)
-        height = round(float(xml_gcp.findtext('.//{*}height')), 7)
-        gcps.append(utils.GeoControlPoint(pixel, line, longitude, latitude, height))
-    return gcps
-
-
-def format_gcps_for_burst(
-    burst_number: int, burst_n_lines: int, swath_gcps: Iterable[utils.GeoControlPoint]
-) -> Iterable[utils.GeoControlPoint]:
-    """Format geolocation control points for a burst by making line numbers
-    relative to the burst, and removing any GCPs that are outside of the burst.
-
-    Args:
-        burst_number: zero-indexed burst number
-        burst_n_lines: number of lines in the burst
-        swath_gcps: list of geolocation control points for the entire swath
-
-    Returns:
-        List of geolocation control points for a particular burst
-    """
-    gcps = []
-    burst_starting_line = burst_number * burst_n_lines
-    for gcp in swath_gcps:
-        gcps.append(utils.GeoControlPoint(gcp.pixel, gcp.line - burst_starting_line, gcp.lon, gcp.lat, gcp.hgt))
-
-    # TODO: Why does this create incorrect geolocation
-    # relevant_gcps = [gcp for gcp in gcps if 0 <= gcp.line <= burst_n_lines]
-    return gcps
 
 
 def get_burst_annotation_data(zipped_safe_path: str, swath_path: str) -> Iterable:
@@ -132,7 +33,6 @@ def get_burst_annotation_data(zipped_safe_path: str, swath_path: str) -> Iterabl
         burst_shape: numpy-style tuple of burst array size (n rows, n columns)
         burst_offsets: uncompressed byte offsets for the bursts contained within
             a swath
-        burst_windows: row and column ranges for the valid data within a burst
     """
     swath_path = Path(swath_path)
     annotation_path = swath_path.parent.parent / 'annotation' / swath_path.with_suffix('.xml').name
@@ -147,10 +47,7 @@ def get_burst_annotation_data(zipped_safe_path: str, swath_path: str) -> Iterabl
     burst_starts = [int(x.findtext('.//{*}byteOffset')) for x in burst_xmls]
     burst_lengths = burst_starts[1] - burst_starts[0]
     burst_offsets = [utils.Offset(x, x + burst_lengths) for x in burst_starts]
-    burst_windows = [compute_valid_window(i, burst_xml) for i, burst_xml in enumerate(burst_xmls)]
-    swath_gcps = get_gcps_from_xml(xml)
-    burst_gcps = [format_gcps_for_burst(i, n_lines, swath_gcps) for i in range(len(burst_xmls))]
-    return burst_shape, burst_offsets, burst_windows, burst_gcps
+    return burst_shape, burst_offsets
 
 
 def create_xml_metadata(zipped_safe_path: str, zinfo: zipfile.ZipInfo) -> utils.XmlMetadata:
@@ -203,16 +100,18 @@ def create_index(
     """
     slc_name = Path(zipped_safe_path).with_suffix('').name
     tiff_name = Path(zinfo.filename).name
+    swath_name = Path(zinfo.filename).name.split('-')[1].upper()
     annotation_name = Path(zinfo.filename).with_suffix('.xml').name
+
     annotation_offset = [item for item in xml_metadatas if item.name == annotation_name][0].offset
     manifest_offset = [item for item in xml_metadatas if item.name == 'manifest.safe'][0].offset
-    burst_shape, burst_offsets, burst_windows, burst_gcps = get_burst_annotation_data(zipped_safe_path, zinfo.filename)
+    burst_shape, burst_offsets = get_burst_annotation_data(zipped_safe_path, zinfo.filename)
 
     bursts = {}
     indexer = utils.ZipIndexer(zipped_safe_path, tiff_name)
     indexer.create_full_dflidx()
-    for i, (burst_offset, burst_window, burst_gcp) in enumerate(zip(burst_offsets, burst_windows, burst_gcps)):
-        burst_name = create_burst_name(slc_name, zinfo.filename, i)
+    for burst_index, burst_offset in enumerate(burst_offsets):
+        burst_name = create_burst_name(slc_name, zinfo.filename, burst_index)
 
         compressed_offset, uncompressed_offset, modified_index = indexer.subset_dflidx(
             locations=[burst_offset.start], end_location=burst_offset.stop
@@ -226,13 +125,13 @@ def create_index(
         burst = utils.BurstMetadata(
             burst_name,
             slc_name,
+            swath_name,
+            burst_index,
             burst_shape,
             compressed_offset,
             index_burst_offset,
             annotation_offset,
             manifest_offset,
-            burst_window,
-            burst_gcp,
         )
 
         if output_json:
